html/dom/documents/resource-metadata-management/document-promises.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>document.parsed, document.contentLoaded, and document.loaded</title>
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-parsed">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-contentloaded">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-document-loaded">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

let onDomContentLoadedFired = false;
let onLoadFired = false;
const readyStateChanges = [];

let parsedFulfilled = false;
let contentLoadedFulfilled = false;
let loadedFulfilled = false;

let parsedAssertions, contentLoadedAssertions, loadedAssertions;

test(() => {
  document.addEventListener("readystatechange", () => {
    readyStateChanges.push(document.readyState);
  });

  document.addEventListener("DOMContentLoaded", () => {
    onDomContentLoadedFired = true;
  });

  window.addEventListener("load", () => {
    onLoadFired = true;
  });

  // A note on timing:
  // The readystatechange event handler function must execute before the promise fulfillment callback does, because the
  // sequence goes:
  // - UA code resolves the promise, thus enqueuing a microtask to call the onFulfilled
  // - UA code fires an event
  //   - Eventually this uses Web IDL to invoke the callback
  //     (https://heycam.github.io/webidl/#call-a-user-objects-operation)
  //   - *After* the callback runs, we perform a microtask checkpoint
  // - The microtask checkpoint calls onFulfilled

  parsedAssertions = document.parsed.then(value => {
    parsedFulfilled = true;

    assert_equals(value, undefined, "The document.parsed promise must fulfill with undefined");
    assert_array_equals(readyStateChanges, ["parsed"],
      "Inside the document.parsed fulfillment handler, the readystatechange event must have fired once already");
    assert_equals(document.readyState, "interactive",
      "Inside the document.parsed fulfillment handler, readyState must be interactive");
  });

  contentLoadedAssertions = document.contentLoaded.then(value => {
    contentLoadedFulfilled = true;

    assert_equals(value, undefined, "The document.contentLoaded promise must fulfill with undefined");
    assert_equals(document.readyState, "interactive",
      "Inside the document.contentLoaded fulfillment handler, readyState must be interactive");
    assert_equals(parsedFulfilled, true,
      "Inside the document.contentLoaded fulfillment handler, document.parsed must have already fulfilled");
    assert_equals(onDomContentLoadedFired, true,
      "Inside the document.contentLoaded fulfillment handler, DOMContentLoaded must have already fired");
    assert_array_equals(readyStateChanges, ["interactive"],
      "Inside the document.contentLoaded fulfillment handler, the readystatechange event must have fired exactly once");
  });

  loadedAssertions = document.loaded.then(value => {
    loadedFulfilled = true;

    assert_equals(value, undefined, "The document.loaded promise must fulfill with undefined");
    assert_equals(document.readyState, "complete",
      "Inside the document.loaded fulfillment handler, readyState must be complete");
    assert_equals(parsedFulfilled, true,
      "Inside the document.loaded fulfillment handler, document.parsed must have already fulfilled");
    assert_equals(contentLoadedFulfilled, true,
      "Inside the document.loaded fulfillment handler, document.contentLoaded must have already fulfilled");
    assert_equals(onDomContentLoadedFired, true,
      "Inside the document.loaded fulfillment handler, DOMContentLoaded must have already fired");
    assert_equals(onLoadFired, false,
      "Inside the document.loaded fulfillment handler, load must not have already fired");
    assert_array_equals(readyStateChanges, ["interactive", "complete"],
      "Inside the document.loaded fulfillment handler, the readystatechange event must have fired exactly twice");
  });
}, "Setup code");

promise_test(() => parsedAssertions || assert_unreached(), "document.parsed");
promise_test(() => contentLoadedAssertions || assert_unreached(), "document.contentLoaded");
promise_test(() => loadedAssertions || assert_unreached(), "document.loaded");

// Historical names must not be supported
test(() => {
  assert_false('interactive' in document);
}, 'document.interactive must not be supported')
</script>
